requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-redis==5.4.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyJWT==2.9.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
sentry-sdk==2.22.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
